Log in screen visual refresh
Also as part of the ongoing efforts to improve user acceptance of the log in process:

- "ID Number" text field label changed to "Username"
- "Sign in" button text changed to "Log In"

because those are the text labels used on the OurVLE website's log in
form. I'm hoping that this consistency will in some way be beneficial.

Furthermore, the colorControlHighlight style item was removed because I
mistakenly thought it was also necessary for styling text fields' colour
in their pressed state.