Remove unnecessary TabLayout style stuff
Most of the TabLayout styling provided was unnecessary and it was
overriding style attributes provided by the Design Support Library's
framework in such a way that the tab layout did not follow the Material Design
Guidelines.

The Design Support Library, however, uses the accent colour by default for the tab
indicator colour so that overriden attribute was left in place.